requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
gradio==4.19.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
scikit-learn==1.4.0
httpx==0.26.0
python-multipart==0.0.6
